usb: api: Clean up for inclusion path. API rework.
By this commit public and private API is reworked.
From user perspective unnecessary things are hidden
by moving them from usb_device.h -> usb_dc.h.

Endpoint related callbacks, structures, etc shall not
be visuble to end user and thus are placed in device
controller header file (usb_dc.h).

Inclusion path is changed so that usb_device.h (public API)
is no longer including (usb_dc.h).

Right now, if user want to implement his own class
or interact direclty with endpoints he needs to include
usb_dc.h.

This commit is solution for #20236.
